메인현황판 전국 데이터 숫자타입 변경 #49 -> int 타입에서 long 타입으로 변경

+++ Ami/src/main/java/com/cnu/ami/dashboard/dao/document/UseDayHourTemp.java
@@ -9,6 +9,6 @@ public class UseDayHourTemp {
 
 	private String day;
 	private int hour;
-	private int sum;
+	private long sum;
 
 }
